Move reusable components into subdirectory
